requirements.txt
Django>=5.0.0
djangorestframework>=3.14.0
django-cors-headers>=4.0.0
